pyproject.toml
[project]
name = "pim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "opendartreader>=0.2.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pykrx>=1.0.51",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.30.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "storage*"]
